+++ panda-market/src/f_shared/lib/index.ts
@@ -3,3 +3,4 @@ export * from './close';
 export * from './dateForm';
 export * from './modal';
 export * from './day';
+export * from './dropdown';
