+++ panda-market/src/f_shared/ui/dropdown/index.ts
@@ -1,2 +1,1 @@
-export * from './lib';
-export * from './ui';
+export * from './Dropdown';
