--- panda-market/src/f_shared/ui/dropdown/lib/useDropdown.ts
@@ -1,1 +1,0 @@
-export const useDropdown = () => {};
