--- panda-market/src/f_shared/ui/dropdown/ui/index.ts
@@ -1,1 +1,0 @@
-export * from './Dropdown';
